Consumer/requirements.txt
pika
python-json-logger
pyjwt
python-dotenv
psycopg2-binary